Create monitoring namespace
Move the monitoring tools (Prometheus, Kiali, Grafana) into their own
namespace. The Kiali operator is set up in its own namespace.  This is
not strictly necessary but it does prevent some annoying warnings.

The ultimate goal is to add the monitoring workspace to the
mesh. However, the current version of Prometheus hangs when started
with the Envoy proxy, so for now the tools are not in the mesh.