Use latest release on GCP "regular" channel
The old commit required a specific Kubernetes release on GCP
clusters. This has already gone out of support on the specified zone.

This commit simply requires the latest release on the GCP "regular"
channel for the specified zone.  This is the most recent release in
General Availability.

The version and channel are now parameterized via the VER_CHAN
environment variable, allowing the user to override via `make -e`.