Add Gatling script to call all S1 and S2 routes
The new scripts AllMusicCalls and AllUserCalls demonstrate how to call
every active route in services S1 and S2.  These calls demonstrate a
range of call types. They also demonstrate how to retain the result of
one call, such as the UUID of an added User, and use that value to
read back the added value.