Shorten Gatling output considerably
This commit removes the instance number from the request type. This
means every request for Music or User is considered the same type of
request and thus are counted together rather than individually. This
dramatically reduces the number of lines of output, which used to grow
with each report interval. Now the lines / report interval are
constant.