Add Jaeger resident memory panel to Flask dashboard
This panel isn't related to the rest of the panels, which describe the
Python Flask attributes. But for someone debugging Jaeger's
performance, it is useful to directly correlate Jaeger's memory with
the performance of the microservices, which will be sharing a node
with the memory-hungry Jaeger process.

If you're only interested in Flask you can comfortably ignore this
panel

Possible alternatives:

1. Separate Grafana dashboard, perhaps with panels for other memory-hungry
processes, such as Prometheus, Gatling, and the overall memory on the
Nodes.

2. Grafana dashboard providing multiple views of Jaeger performance.